% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cwsdb.R
\name{build_cwsdb}
\alias{build_cwsdb}
\title{build_cwsdb}
\usage{
build_cwsdb()
}
\value{
cwsdb
}
\description{
Build the city water supply database
}
\details{
Builds the cws entry for Tokyo
}
\author{
SWDT March 2018
}
