% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_delhi.R, R/build_tempate_city.R
\name{build_delhi}
\alias{build_delhi}
\alias{build_delhi}
\title{build_delhi}
\usage{
build_delhi()

build_delhi()
}
\value{
A row of gluwasp data for the city of Delhi, India

A row of gluwasp data for the city of XXXX, COUNTRY XXXX
}
\description{
Build the Delhi gluwasp entry from raw data

Build the XXXX gluwasp entry from raw data
}
\details{
Builds the gluwasp entry for Delhi

Builds the gluwasp entry for XXXX
}
\author{
SWDT April 2018

SWDT April 2018
}
