% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gluwasp.R
\name{build_gluwasp}
\alias{build_gluwasp}
\title{build_gluwasp}
\usage{
build_gluwasp()
}
\value{
gluwasp
}
\description{
Build the city water supply database
}
\details{
Builds the gluwasp database
}
\author{
SWDT May 2018
}
