% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tokyo.R
\name{build_tokyo}
\alias{build_tokyo}
\title{build_tokyo}
\usage{
build_tokyo()
}
\value{
A row of gluwasp data for the city of Tokyo, Japan.
}
\description{
Build the Tokyo gluwasp entry from raw data
}
\details{
Builds the gluwasp entry for Tokyo
}
\author{
SWDT April 2018
}
