% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwsdb-package.R
\docType{package}
\name{cwsdb-package}
\alias{cwsdb}
\alias{cwsdb-package}
\title{cwsdb: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sean Turner \email{sean.turner@protonmail.com} (0000-0003-4400-9800)

}
\keyword{internal}
