% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tokyo.R
\name{build_tokyo}
\alias{build_tokyo}
\title{build_tokyo}
\usage{
build_tokyo()
}
\value{
A row of
}
\description{
Build the Tokyo gluwasp entry from raw data
}
\details{
Builds the cws entry for Tokyo
}
\author{
SWDT March 2018
}
